webkit/data/layout_test_results/v8/LayoutTests/svg/custom/immutable-properties-expected.txt

CONSOLE MESSAGE: line 21: Uncaught TypeError: Cannot set property 'x' of undefined
Tests whether immutable properties can not be modified.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS viewport.x is 0
PASS viewport.x is 100
PASS svgDoc.documentElement.viewport.x is 0
FAIL zoomRectScreen.x should be 0. Threw exception TypeError: Cannot read property 'x' of undefined
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
